backend/requirements.txt
Django==3.2
djangorestframework
djoser
# django-extra-fields
pillow==10.0.0